arm: Switch to reboot
The ARMv5 and ARMv6 machines don't know how to exit the guest on
shutdown. Instead use Qemu's -no-reboot option along with a modified
rootfs that calls 'reboot' instead of 'shutdown'.

Note that the ARMv7 machine is also moved over to this method so we can
continue sharing an the root file system.

Signed-off-by: Joel Stanley <joel@jms.id.au>