Rework for aspeed testing
Remove the existing configurations and add variants for testing aspeed
images.

The kernel trees tested are:

 - openbmc's -dev trees
 - linus master
 - linux-next

The kernel config and machines tested on are:
 - ast2600evb / aspeed_g5_defconfig
 - romulus-bmc / aspeed_g5_defconfig
 - palmetto-bmc / aspeed_g4_defconfig

The mutli-machine kernel configs and machines tested on are:

 - ast2600evb / multi_v7_defconfig
 - palmetto-bmc / multi_v5_defconfig

There is no multi v6 configuration.

This removes the dependency on boot-util's boot-qemu.sh so the qemu
command lines can be customed without having to maintain a fork. The
downside is it adds back in the mix of building and booting Linux to
driver.sh.

Signed-off-by: Joel Stanley <joel@jms.id.au>